assignment authorization

+++ src/main/java/springbeam/susukgwan/assignment/AssignmentService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import springbeam.susukgwan.ResponseMsg;
 import springbeam.susukgwan.ResponseMsgList;
@@ -15,6 +16,7 @@ import springbeam.susukgwan.note.NoteRepository;
 import springbeam.susukgwan.tutoring.Tutoring;
 import springbeam.susukgwan.tutoring.TutoringRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -32,8 +34,16 @@ public class AssignmentService {
     /* 숙제 추가 */
     public ResponseEntity<?> createAssignment(AssignmentRequestDTO.Create createAssignment) {
 
+        /* Authorization - 이 수업의 선생님인 유저만 접근 가능 */
+        Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+
         Optional<Tutoring> tutoring = tutoringRepository.findById(createAssignment.getTutoringId());
         if (tutoring.isPresent()) {
+            if (tutoring.get().getTutorId() != userId) { // Tutoring에 등록된 선생님과 현재유저(선생님)가 다르면 불가능
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMsg(ResponseMsgList.NOT_AUTHORIZED.getMsg()));
+            }
+        /* End */
+
             Optional<Note> note = noteRepository.findFirst1ByTutoringOrderByDateTimeDesc(tutoring.get());
             if (note.isPresent()) {
                 Assignment assignment = Assignment.builder()
@@ -57,6 +67,15 @@ public class AssignmentService {
 
     /* 숙제 수정 */
     public ResponseEntity<?> updateAssignment(Long assignmentId, AssignmentRequestDTO.Update updateAssignment) {
+
+        /* Authorization - 이 숙제를 등록한 선생님만 접근 가능 */
+        Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+        Long tutorIdOfAssignment = assignmentRepository.GetTutorIdOfAssignment(assignmentId);
+        if (tutorIdOfAssignment != null && userId != tutorIdOfAssignment) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMsg(ResponseMsgList.NOT_AUTHORIZED.getMsg()));
+        }
+        /* End */
+
         Optional<Assignment> assignment = assignmentRepository.findById(assignmentId);
 
         if (assignment.isEmpty()) {
@@ -94,6 +113,15 @@ public class AssignmentService {
 
     /* 숙제 삭제 */
     public ResponseEntity<?> deleteAssignment(Long assignmentId) {
+
+        /* Authorization - 이 숙제를 등록한 선생님만 접근 가능 */
+        Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+        Long tutorIdOfAssignment = assignmentRepository.GetTutorIdOfAssignment(assignmentId);
+        if (tutorIdOfAssignment != null && userId != tutorIdOfAssignment) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMsg(ResponseMsgList.NOT_AUTHORIZED.getMsg()));
+        }
+        /* End */
+
         Optional<Assignment> assignment = assignmentRepository.findById(assignmentId);
         if (assignment.isPresent()) {
             assignmentRepository.deleteById(assignmentId);
@@ -105,6 +133,20 @@ public class AssignmentService {
 
     /* 숙제의 모든 인증피드 리스트 */
     public ResponseEntity<?> submitListOfAssignment (Long assignmentId) {
+
+        /* Authorization - 이 숙제에 할당된 선생님, 학생, 학부모만 접근 가능 */
+        Long userId = Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
+
+        List<Long> allUserIdOfAssignment = new ArrayList<>();
+        allUserIdOfAssignment.add(assignmentRepository.GetTutorIdOfAssignment(assignmentId));
+        allUserIdOfAssignment.add(assignmentRepository.GetTuteeIdOfAssignment(assignmentId));
+        allUserIdOfAssignment.add(assignmentRepository.GetParentIdOfAssignment(assignmentId));
+
+        if (!allUserIdOfAssignment.isEmpty() && !allUserIdOfAssignment.contains(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(new ResponseMsg(ResponseMsgList.NOT_AUTHORIZED.getMsg()));
+        }
+        /* End */
+
         Optional<Assignment> assignment = assignmentRepository.findById(assignmentId);
 
         if (assignment.isEmpty()) {
